Implemented player profile system
