+++ .gitignore
@@ -3,3 +3,5 @@ __pycache__
 *.code-workspace
 settings.json
 TESTENV.py
+[0-9]*.json
+_*.json
