+++ main.py
@@ -2,6 +2,7 @@ import discord, json, os
 
 from discordlib import *
 from dicelib import *
+from characterlib import get_character
 from pathlib import Path
 from discord.ext import commands
 
@@ -9,76 +10,17 @@ class Struct:
     def __init__(self, data):
         self.__dict__.update(data)
 
-class Character:
-    def __init__(self, **data):
-        self.player_id = data["player_id"]
-        
-        self.name = data["name"]
-        self.job = data["job"]
-        self.level = data["level"]
-
-        self.strength = data["strength"]
-        self.dexterity = data["dexterity"]
-        self.constitution = data["constitution"]
-        self.wisdom = data["wisdom"]
-        self.intelligence = data["intelligence"]
-        self.charisma = data["charisma"]
-
-        self.proficiencies = data["proficiencies"]
-    
-    def get_proficiency(self, skill: str) -> int:
-        if skill not in self.proficiencies:
-            return 0
-        return self.proficiencies[skill]
-    
-    def edit_data(self, key: str, val: str|int) -> None:
-        if hasattr(self, key):
-            if type(getattr(self, key, False)) != type(val):
-                return
-            setattr(self, key, val)
-    
-    def edit_proficiency(self, skill: str, val: int) -> None:
-        if val == 0:
-            self.proficiencies.pop(skill, None)
-            return
-            
-        self.proficiencies[skill] = val
-    
-    def save(self):
-        save_file = os.path.join(os.path.dirname(__file__), f"data/profiles/{self.player_id}.json")
-        mode = "r+" if Path(save_file) else "w+"
-        
-        with open(save_file, mode) as f:
-            profile = json.loads(f.read())
-
-            if self.name not in profile.characters:
-                profile.characters[self.name] = {}
-            
-            profile.characters[self.name].update(self.__dict__)
-
-            f.write(json.dumps(profile))      
-
-
 with open(os.path.join(os.path.dirname(__file__), "settings.json")) as f:
     settings = Struct(json.loads(f.read()))
 
 intents = discord.Intents.all()
 bot = commands.Bot(command_prefix="!", intents=intents)
-
 guild = None # set at on_ready event
 
 save_file = os.path.join(os.path.dirname(__file__), "data/main.json")
 
 # Serialized data
 dungeon_master_id = None
-
-"""
-player_id: {
-    current_character: str,
-    characters: dict
-}
-"""
-characters = {}
 
 
 def save_data():
@@ -108,39 +50,6 @@ def get_member(id: int) -> discord.Member:
     """Returns a discord.Member object from the guild using a user id."""
     return guild.get_member(int(id))
 
-
-def profile_exists(id: int) -> str:
-    return Path(f"data/profiles/{id}.json").exists()
-
-
-def load_character(id: int, character: str = None):
-    if not profile_exists(id):
-        return
-
-    with open(f"data/profiles/{id}.json", "r+"):
-        f_data = f.read()
-        if f_data == "": return
-        data = json.loads(f_data)
-    
-        character = character if character else data["current_character"]
-        if not character:
-            return
-        
-        characters[id] = Character(**data["characters"][character])
-        
-        data["current_character"] = character
-        f.write(json.dumps(data))
-
-
-def get_character(id: int) -> Character:
-    if id not in characters:
-        load_character(id)
-    
-    if id in characters:
-        return characters[id]
-    
-    return {}
-    
 
 @bot.event
 async def on_ready():
@@ -176,13 +85,15 @@ async def register_dm(ctx, user : discord.Member = None):
 
 @bot.command(name="roll")
 async def roll_dice(ctx, op1 = None, op2 = None, op3 = None):
-    string, embed = get_roll_results(op1, op2, op3)
+    character = get_character(ctx.author.id)
+    string, embed = get_roll_results(character, op1, op2, op3)
     await send_message(ctx.channel, string, embed)
 
 
 @bot.command(name="rolldm")
 async def roll_dice_dm(ctx, op1 = None, op2 = None, op3 = None):
-    string, embed = get_roll_results(op1, op2, op3)
+    character = get_character(ctx.author.id)
+    string, embed = get_roll_results(character, op1, op2, op3)
     dm = get_member(dungeon_master_id)
 
     if ctx.author != dm or not embed:
